+++ src/wildtornado/scocalc/strategies/LinesOfCodeCalculator.java
@@ -2,12 +2,12 @@ package wildtornado.scocalc.strategies;
 
 public class LinesOfCodeCalculator extends BaseCalculator {
 
-    public LinesOfCodeCalculator(int val) {
+    public LinesOfCodeCalculator(double val) {
         this.val = val;
     }
 
     public void generate() {
-        this.result = val;
+        this.result = (int) val;
     }
 
 }
